fix: native search sdk 1.1.0 update for startup fixes
