+++ MapboxSearch/autofill/src/main/java/com/mapbox/search/autofill/AddressAutofill.kt
@@ -6,6 +6,7 @@ import com.mapbox.android.core.location.LocationEngineProvider
 import com.mapbox.bindgen.Expected
 import com.mapbox.geojson.Point
 import com.mapbox.search.base.BaseSearchSdkInitializer
+import com.mapbox.search.base.BaseSearchSdkInitializerImpl
 import com.mapbox.search.base.location.defaultLocationEngine
 
 /**
@@ -75,7 +76,7 @@ public interface AddressAutofill {
         ): AddressAutofill {
             return AddressAutofillImpl.create(
                 accessToken = accessToken,
-                app = BaseSearchSdkInitializer.app,
+                app = BaseSearchSdkInitializerImpl.app,
                 locationEngine = locationEngine
             )
         }
