+++ MapboxSearch/place-autocomplete/src/main/java/com/mapbox/search/autocomplete/PlaceAutocomplete.kt
@@ -7,6 +7,7 @@ import com.mapbox.bindgen.Expected
 import com.mapbox.geojson.BoundingBox
 import com.mapbox.geojson.Point
 import com.mapbox.search.base.BaseSearchSdkInitializer
+import com.mapbox.search.base.BaseSearchSdkInitializerImpl
 import com.mapbox.search.base.location.defaultLocationEngine
 
 /**
@@ -81,7 +82,7 @@ public interface PlaceAutocomplete {
         ): PlaceAutocomplete {
             return PlaceAutocompleteImpl.create(
                 accessToken = accessToken,
-                app = BaseSearchSdkInitializer.app,
+                app = BaseSearchSdkInitializerImpl.app,
                 locationEngine = locationEngine
             )
         }
