+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/MapboxSearchSdkInitializer.kt
@@ -1,17 +1,7 @@
 package com.mapbox.search
 
-import android.app.Application
-import android.content.Context
-import androidx.startup.Initializer
-import com.mapbox.search.base.BaseSearchSdkInitializer
+import com.mapbox.common.BaseMapboxInitializer
 
-internal class MapboxSearchSdkInitializer : Initializer<MapboxSearchSdk> {
-    override fun create(context: Context): MapboxSearchSdk {
-        MapboxSearchSdk.initialize(context.applicationContext as Application)
-        return MapboxSearchSdk
-    }
-
-    override fun dependencies(): List<Class<out Initializer<*>>> {
-        return listOf(BaseSearchSdkInitializer::class.java)
-    }
+internal class MapboxSearchSdkInitializer : BaseMapboxInitializer<MapboxSearchSdk>() {
+    override val initializerClass = MapboxSearchSdkInitializerImpl::class.java
 }
